// movie-frontend-app/src/app/services/models/page-response-movie-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MovieResponse } from '../models/movie-response';
export interface PageResponseMovieResponse {
  content?: Array<MovieResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPage?: number;
}